+++ src/LightOps.Commerce.Gateways.Storefront/Api/Providers/IMetaFieldEndpointProvider.cs
@@ -1,0 +1,8 @@
+﻿namespace LightOps.Commerce.Gateways.Storefront.Api.Providers
+{
+    public interface IMetaFieldEndpointProvider
+    {
+        bool IsEnabled { get; }
+        string GrpcEndpoint { get; }
+    }
+}